Add files via upload
This code is the combination of all the codes in this folder.
It can interpolate, regress, approximate the integral of a particular function, and solve a particular differential equation by taking data from the user. It performs each of these operations using different methods and compares them with each other.